structs, methods & interfaces/shapes_test.go
package main

import (
	"testing"
)

func TestPerimeter(t *testing.T) {
	triangle := Triangle{10.0, 10.0, 10.0}
	got := Perimeter(triangle)
	want := 30.0

	if got != want {
		t.Errorf("got %.2f want %.2f", got, want)
	}
}

// func TestArea(t *testing.T) {
// 	checkArea := func(t testing.TB, shape Shape, want float64) {
// 		t.Helper()
// 		got := shape.Area()
// 		if got != want {
// 			t.Errorf("got %g want %g", got, want)
// 		}
// 	}

// 	t.Run("triangle", func(t *testing.T) {
// 		triangle := Triangle{10.517, 6.0, 7.0}
// 		checkArea(t, triangle, 20.0)
// 	})

// 	t.Run("circles", func(t *testing.T) {
// 		circle := Circle{10}
// 		checkArea(t, circle, 314.1592653589793)
// 	})
// }

func TestArea(t *testing.T) {

	areaTests := []struct {
		shape Shape
		want  float64
	}{
		{Triangle{10.517, 6.0, 7.0}, 20.0},
		{Circle{10}, 314.1592653589793},
	}

	for _, tt := range areaTests {
		got := tt.shape.Area()
		if got != tt.want {
			t.Errorf("got %g want %g", got, tt.want)
		}
	}

}
